Relocate range utils
