+++ src/op_analytics/datapipeline/etl/ingestion/construct.py
@@ -1,9 +1,9 @@
 from op_analytics.coreutils import clickhouse
 from op_analytics.coreutils.logger import structlog
 from op_analytics.coreutils.partitioned.location import DataLocation
+from op_analytics.coreutils.rangeutils.blockrange import BlockRange
+from op_analytics.coreutils.rangeutils.timerange import TimeRange
 from op_analytics.coreutils.threads import run_concurrently
-from op_analytics.datapipeline.utils.blockrange import BlockRange
-from op_analytics.datapipeline.utils.timerange import TimeRange
 
 from .batches import BlockBatch, split_block_range
 from .sources import RawOnchainDataProvider
